genai/pyproject.toml
[project]
name = "genai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "strands-agents>=1.4.0",
    "strands-agents-tools>=0.2.3",
    "bedrock-agentcore>=0.1.1",
    "bedrock-agentcore-starter-toolkit>=0.1.0",
    "uvicorn>=0.35.0",
    "pandas>=2.0.0",
    "boto3>=1.34.0",
    "requests>=2.32.4",
]
